backend/requirements.txt
alembic==1.7.5
bidict==0.22.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.3
cryptography==38.0.1
firebase-admin==6.0.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==2.5.1
gevent==22.10.2
gevent-websocket==0.10.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==2.0.1
grpcio==1.50.0
grpcio-status==1.50.0
gunicorn==20.1.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
msgpack==1.0.4
mysqlclient==2.1.0
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.19.2
python-engineio==4.3.4
python-socketio==5.7.2
ref==0.0.2.2
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.28
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.0.2
wsproto==1.2.0
zope.event==4.6
zope.interface==5.5.2
